+++ src/visualization/SingleVolumeBrickVT.cpp
@@ -50,7 +50,7 @@ void SingleVolumeBrickVT::prepare(const Scene& scene)
         struct Edit {
             uint32    type;
             uint32    op;
-            uint32    blending;
+            float32   blending;
             uint32    _padding1;
             glm::vec3 pos;
             float     _padding2;
